Add workspace location to workspace preferences
In Preferences > General > Workspace we added
workspace location label and Text box which
shows the path to the current workspace.
This was done similar to the way Resource Info
dialog does that.

This is the first part of the enhancement requested in:
Bug 432480 - [Workbench] [Preferences] Preferences -> General -> Workspace should show workspace location and have option show Workspace location

Bug: 432480